recipes encapsulation, new folder for recipes, rename paths for update recipe and Show recipes. This is a save previous to other modifications also is necesary modify this 3 files to validate push complete recipe whit image
