+++ reactfront/src/config/routes/paths.js
@@ -12,6 +12,7 @@ export const USER  = '/private/user'
 export const PRUEBA  = '/private/prueba'
 export const PRUEBA2  = '/private/prueba2'
 export const POSTR  = '/private/post'
+export const UPDATER  = '/private/postUpdates'
 export const SHOWPOSTR  = '/private/postShowRecipe'
 export const UPLOADTEST = '/private/uploadTest'
